+++ lib/definitions/redefiner.rb
@@ -11,8 +11,13 @@ module LowType
   class Redefiner
     class << self
       def redefine(method_nodes:, class_proxy:, file_path:)
-        method_proxies = create_method_proxies(method_nodes:, klass: class_proxy.klass, file_path:)
-        define_methods(method_proxies:, class_proxy:)
+        method_proxies = build_methods(method_nodes:, klass: class_proxy.klass, file_path:)
+
+        if LowType.config.type_checking
+          typed_methods(method_proxies:, class_proxy:)
+        else
+          untyped_methods(method_proxies:, class_proxy:)
+        end
       end
 
       def redefinable?(method_proxy:, class_proxy:)
@@ -32,9 +37,24 @@ module LowType
         true
       end
 
+      def untyped_args(args:, kwargs:, method_proxy:) # rubocop:disable Metrics/AbcSize
+        method_proxy.params.each do |param_proxy|
+          value = param_proxy.position ? args[param_proxy.position] : kwargs[param_proxy.name]
+
+          next unless value.nil?
+          raise param_proxy.error_type, param_proxy.error_message(value:) if param_proxy.required?
+
+          value = param_proxy.type_expression.default_value # Default value can still be `nil`.
+          value = value.value if value.is_a?(ValueExpression)
+          param_proxy.position ? args[param_proxy.position] = value : kwargs[param_proxy.name] = value
+        end
+
+        [args, kwargs]
+      end
+
       private
 
-      def create_method_proxies(method_nodes:, klass:, file_path:)
+      def build_methods(method_nodes:, klass:, file_path:)
         method_nodes.each do |name, method_node|
           file = ProxyFactory.file_proxy(path: file_path, node: method_node, scope: "#{klass}##{name}")
 
@@ -48,21 +68,19 @@ module LowType
         Repository.all(klass:)
       end
 
-      def define_methods(method_proxies:, class_proxy:) # rubocop:disable Metrics
+      def typed_methods(method_proxies:, class_proxy:) # rubocop:disable Metrics
         Module.new do
           method_proxies.each do |name, method_proxy|
             next unless LowType::Redefiner.redefinable?(method_proxy:, class_proxy:)
 
-            # NOTE: You are now in the binding of the includer class (`name` is also available here).
+            # You are now in the binding of the includer class (`name` is also available here).
             define_method(name) do |*args, **kwargs|
               # Inlined version of Repository.load() for performance increase.
               method_proxy = instance_of?(Class) ? low_methods[name] : self.class.low_methods[name] || Object.low_methods[name]
 
               method_proxy.params.each do |param_proxy|
                 value = param_proxy.position ? args[param_proxy.position] : kwargs[param_proxy.name]
-                if value.nil? && param_proxy.type_expression.default_value != :LOW_TYPE_UNDEFINED
-                  value = param_proxy.type_expression.default_value
-                end
+                value = param_proxy.type_expression.default_value if value.nil? && !param_proxy.required?
 
                 param_proxy.type_expression.validate!(value:, proxy: param_proxy)
                 value = value.value if value.is_a?(ValueExpression)
@@ -75,6 +93,24 @@ module LowType
                 return return_value
               end
 
+              super(*args, **kwargs)
+            end
+
+            private name if class_proxy.private_start_line && method_proxy.start_line > class_proxy.private_start_line
+          end
+        end
+      end
+
+      def untyped_methods(method_proxies:, class_proxy:)
+        Module.new do
+          method_proxies.each do |name, method_proxy|
+            next unless LowType::Redefiner.redefinable?(method_proxy:, class_proxy:)
+
+            # You are now in the binding of the includer class (`name` is also available here).
+            define_method(name) do |*args, **kwargs|
+              # NOTE: Type checking is currently disabled. See 'config.type_checking'.
+              method_proxy = instance_of?(Class) ? low_methods[name] : self.class.low_methods[name] || Object.low_methods[name]
+              args, kwargs = LowType::Redefiner.untyped_args(args:, kwargs:, method_proxy:)
               super(*args, **kwargs)
             end